Refactoring for Tiers Structuring
Added .gitignore, split functionality of the actually draft running into Draft.py. Tested lists, randoms, picks, and swaps.
